+++ Addon/ElvUI_mMediaTag/modules/misc/Tags.lua
@@ -1858,6 +1858,7 @@ local targetTextures = {
 
 local targetStringColors = {}
 local function GetPartyTargetsIcons(unit, style)
+	if not  mMT.Classes then return end
 	if not targetStringColors.build then
 		for _, class in ipairs(mMT.Classes) do
 			local color = E:ClassColor(class)
@@ -1964,6 +1965,11 @@ local classIcons = {
 	outlineplus = "mmt_transparent_outline_colorboost.tga",
 	outlineshadow = "mmt_transparent_outline_shadow.tga",
 	outlineshadowplus = "mmt_transparent_outline_shadow_colorboost.tga",
+	simple = "mmt_simple.tga",
+	simpleborder = "mmt_simple_border.tga",
+	simpleclass = "mmt_simple_class.tga",
+	simpletrasnparent = "mmt_simple_trans.tga",
+	simpleclassbg = "mmt_simple_class_bg.tga",
 }
 
 for style, file in next, classIcons do